+++ Sources/LocaleSupport/Locale+Identifiable.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+#if hasFeature(RetroactiveAttribute)
+extension Locale: @retroactive Identifiable {
+    public var id: String { identifier }
+}
+#else
+extension Locale: Identifiable {
+    public var id: String { identifier }
+}
+#endif
